+++ src/world/generator/executor/AsyncGeneratorUnregisterTask.php
@@ -21,30 +21,22 @@
 
 declare(strict_types=1);
 
-namespace pocketmine\world\generator;
+namespace pocketmine\world\generator\executor;
 
 use pocketmine\scheduler\AsyncTask;
 use pocketmine\world\World;
 
-class GeneratorRegisterTask extends AsyncTask{
+/**
+ * @internal
+ */
+class AsyncGeneratorUnregisterTask extends AsyncTask{
 	public int $worldId;
-	public int $worldMinY;
-	public int $worldMaxY;
 
-	/**
-	 * @phpstan-param \Closure() : Generator $generatorFactory
-	 */
-	public function __construct(
-		World $world,
-		private readonly \Closure $generatorFactory
-	){
+	public function __construct(World $world){
 		$this->worldId = $world->getId();
-		$this->worldMinY = $world->getMinY();
-		$this->worldMaxY = $world->getMaxY();
 	}
 
 	public function onRun() : void{
-		$generator = ($this->generatorFactory)();
-		ThreadLocalGeneratorContext::register(new ThreadLocalGeneratorContext($generator, $this->worldMinY, $this->worldMaxY), $this->worldId);
+		ThreadLocalGeneratorContext::unregister($this->worldId);
 	}
 }
